AG-10836 - Move animation out of candleStickSeriesBase
