+++ packages/ag-charts-enterprise/src/series/candlestick/candlestickSeriesBase.ts
@@ -8,15 +8,9 @@ import {
     _Util,
 } from 'ag-charts-community';
 
-import { type CandlestickBaseGroup, GroupTags } from './candlestickGroup';
+import type { CandlestickBaseGroup } from './candlestickGroup';
 import type { CandlestickSeriesBaseItems, CandlestickSeriesBaseProperties } from './candlestickSeriesProperties';
 import type { CandlestickNodeBaseDatum } from './candlestickTypes';
-import {
-    prepareCandlestickBodyAnimationFunctions,
-    prepareCandlestickWickAnimationFunctions,
-    resetCandlestickSelectionsStartFn,
-    resetCandlestickWickSelectionsStartFn,
-} from './candlestickUtil';
 
 const {
     extent,
@@ -32,7 +26,6 @@ const {
     mergeDefaults,
     isFiniteNumber,
 } = _ModuleSupport;
-const { motion } = _Scene;
 
 const { sanitizeHtml, Logger } = _Util;
 const { ContinuousScale, OrdinalTimeScale } = _Scale;
@@ -364,58 +357,6 @@ export abstract class CandlestickSeriesBase<
         );
     }
 
-    protected override animateEmptyUpdateReady({
-        datumSelection,
-    }: _ModuleSupport.CartesianAnimationData<CandlestickBaseGroup<TNodeDatum, TItemOptions>, TNodeDatum>) {
-        this.resetAnimations(datumSelection);
-    }
-
-    protected override animateReadyResize({
-        datumSelection,
-    }: _ModuleSupport.CartesianAnimationData<CandlestickBaseGroup<TNodeDatum, TItemOptions>, TNodeDatum>) {
-        this.resetAnimations(datumSelection);
-    }
-
-    private resetAnimations(
-        datumSelection: _Scene.Selection<CandlestickBaseGroup<TNodeDatum, TItemOptions>, TNodeDatum>
-    ) {
-        const rects = datumSelection.selectByTag<_Scene.Rect>(GroupTags.Rect);
-        const wicks = datumSelection.selectByClass<_Scene.Line>(_Scene.Line);
-
-        motion.resetMotion(rects, resetCandlestickSelectionsStartFn());
-        motion.resetMotion(wicks, resetCandlestickWickSelectionsStartFn());
-    }
-
-    protected override animateWaitingUpdateReady({
-        datumSelection,
-    }: _ModuleSupport.CartesianAnimationData<CandlestickBaseGroup<TNodeDatum, TItemOptions>, TNodeDatum>) {
-        const { processedData } = this;
-        const difference = processedData?.reduced?.diff;
-        const rects = datumSelection.selectByTag<_Scene.Rect>(GroupTags.Rect);
-        const wicks = datumSelection.selectByClass<_Scene.Line>(_Scene.Line);
-
-        const bodyAnimationFns = prepareCandlestickBodyAnimationFunctions();
-        motion.fromToMotion(
-            this.id,
-            'datums',
-            this.ctx.animationManager,
-            rects,
-            bodyAnimationFns,
-            (_, datum) => String(datum.xValue),
-            difference
-        );
-
-        const wickAnimationFns = prepareCandlestickWickAnimationFunctions();
-        motion.fromToMotion(
-            this.id,
-            'datums',
-            this.ctx.animationManager,
-            wicks,
-            wickAnimationFns,
-            (_, datum) => String(datum.xValue),
-            difference
-        );
-    }
     protected override isVertical(): boolean {
         return true;
     }
@@ -491,6 +432,9 @@ export abstract class CandlestickSeriesBase<
         return this.getSeriesStyles(nodeDatum);
     }
 
+    protected abstract resetAnimations(
+        datumSelection: _Scene.Selection<CandlestickBaseGroup<TNodeDatum, TItemOptions>, TNodeDatum>
+    ): void;
     protected abstract getFormatterParams(params: AgCandlestickSeriesBaseFormatterParams<TNodeDatum>): TFormatterParams;
     protected abstract getSeriesStyles(_nodeDatum: TNodeDatum): TItemOptions;
     protected abstract getActiveStyles(nodeDatum: TNodeDatum, highlighted: boolean): TItemOptions;
